feat/graphql 개발 셋팅 (#10)
우선순위 1 - graphql 개발 셋팅 #7에 대한 작업 완료
