+++ src/api/movie/movie.graphql
@@ -1,0 +1,8 @@
+type Movie {
+  title: String
+  releaseDate: String
+}
+
+type Query {
+  movies: [Movie]
+}
