+++ src/api/movie/movie.ts
@@ -1,0 +1,7 @@
+const resolvers = {
+  Query: {
+    movies: () => ([{title: '코딩 사나이', releaseDate: '2021-08-22'}]),
+  },
+};
+
+export default resolvers;
