+++ src/api/tv/tv.graphql
@@ -1,0 +1,8 @@
+type Tv {
+  title: String
+  releaseDate: String
+}
+
+type Query {
+  tvs: [Tv]
+}
