+++ src/api/tv/tv.ts
@@ -1,0 +1,7 @@
+const resolvers = {
+  Query: {
+    tvs: () => ([{title: '거침없이 코딩 중', releaseDate: '2021-08-22'}]),
+  },
+};
+
+export default resolvers;
