+++ src/typedefs-resolvers/_mutations.ts
@@ -1,0 +1,10 @@
+const { gql } = require("apollo-server");
+
+// const typeDefs = gql`
+//   type Mutation {
+    
+//   }
+// `;
+
+module.exports = typeDefs;
+export {}
