+++ src/typedefs-resolvers/_queries.ts
@@ -1,0 +1,10 @@
+const { gql } = require("apollo-server");
+
+const typeDefs = gql`
+  type Query {
+    books: [Book]
+  }
+`;
+
+module.exports = typeDefs;
+export {}
