+++ tsconfig.json
@@ -1,34 +1,40 @@
 {
-  "compilerOptions": {
-      "outDir": "build/dist",
-      "module": "commonjs",
-      "target": "es6",
-      "lib": [
-          "es6",
-          "dom"
-      ],
-      "sourceMap": true,
-      "allowJs": true,
-      "removeComments": true,
-      "experimentalDecorators": true,
-      "emitDecoratorMetadata": true,
-      "importHelpers": true,
-      "moduleResolution": "node",
-      "rootDir": "src",
-      "noImplicitReturns": true,
-      "noImplicitThis": true,
-      "noImplicitAny": true,
-      "strictNullChecks": true,
-      "esModuleInterop": true,
-      "types": ["node"],
-      "typeRoots": ["../node_modules/@types"]
-  },
-  "exclude": [
-      "node_modules",
-      "build",
-      "scripts",
-      "acceptance-tests",
-      "webpack",
-  ],
-  "include": ["./src/**/*.ts"],
+    "compilerOptions": {
+        "outDir": "build/dist",
+        "module": "commonjs",
+        "target": "es6",
+        "lib": [
+            "es6",
+            "dom",
+        ],
+        "sourceMap": true,
+        "allowJs": true,
+        "removeComments": true,
+        "experimentalDecorators": true,
+        "emitDecoratorMetadata": true,
+        "importHelpers": true,
+        "moduleResolution": "node",
+        "rootDir": "src",
+        "noImplicitReturns": true,
+        "noImplicitThis": true,
+        "noImplicitAny": true,
+        "strictNullChecks": true,
+        "esModuleInterop": true,
+        "types": [
+            "node"
+        ],
+        "typeRoots": [
+            "node_modules/@types"
+        ],
+    },
+    "exclude": [
+        "node_modules",
+        "build",
+        "scripts",
+        "acceptance-tests",
+        "webpack",
+    ],
+    "include": [
+        "./src/**/*.ts"
+    ],
 }
